Adding Fav Icons fetch to Login View Controller
